+++ projects/python/perception/object_detection_2d/nanodet/benchmark_demo.py
@@ -22,22 +22,26 @@ if __name__ == '__main__':
     parser = argparse.ArgumentParser()
     parser.add_argument("--device", help="Device to use (cpu, cuda)", type=str, default="cuda", choices=["cuda", "cpu"])
     parser.add_argument("--model", help="Model for which a config file will be used", type=str, default="m")
+    parser.add_argument("--optimize-jit", help="", default=False, action="store_true")
+    parser.add_argument("--optimize-onnx", help="", default=False, action="store_true")
+    parser.add_argument("--repetitions", help="Determines the amount of repetitions to run", type=int, default=1000)
+    parser.add_argument("--warmup", help="Determines the amount of warmup runs", type=int, default=100)
+    parser.add_argument("--nms", help="Determines the max amount of bboxes the nms will output", type=int, default=30)
     parser.add_argument("--path", help="Path to the image that is used for inference", type=str,
                         default="./predefined_examples/000000000036.jpg")
-    parser.add_argument("--optimize", help="If specified will determine the optimization to be used (onnx, jit)",
-                        type=str, default="", choices=["", "onnx", "jit"])
     args = parser.parse_args()
 
     nanodet = NanodetLearner(model_to_use=args.model, device=args.device)
-    nanodet.download("./predefined_examples", mode="pretrained")
-    nanodet.load("./predefined_examples/nanodet_{}".format(args.model), verbose=True)
-    nanodet.download("./predefined_examples", mode="images")
+    nanodet.download("./predefined_examples", mode="pretrained", verbose=False)
+    nanodet.load("./predefined_examples/nanodet_{}".format(args.model), verbose=False)
+    nanodet.download("./predefined_examples", mode="images", verbose=False)
 
     img = Image.open(args.path)
 
-    if args.optimize != "":
-        nanodet.optimize("./{}/nanodet_{}".format(args.optimize, args.model), optimization=args.optimize)
 
-    boxes = nanodet.infer(input=img, conf_threshold=0.35, iou_threshold=0.6, nms_max_num=20)
+    if args.optimize_jit:
+        nanodet.optimize(f"./jit/nanodet_{args.model}", optimization="jit", verbose=False)
+    if args.optimize_onnx:
+        nanodet.optimize(f"./onnx/nanodet_{args.model}", optimization="onnx", verbose=False)
 
-    draw_bounding_boxes(img.opencv(), boxes, class_names=nanodet.classes, show=True)
+    nanodet.benchmark(img, repetitions=args.repetitions, warmup=args.warmup, nms_max_num=args.nms, half_precision=True)