+++ projects/python/perception/object_detection_2d/nanodet/webcam_demo.py
@@ -18,20 +18,21 @@ import cv2
 import time
 
 from opendr.engine.data import Image
-from opendr.perception.object_detection_2d import YOLOv5DetectorLearner
+from opendr.perception.object_detection_2d import NanodetLearner
 from opendr.perception.object_detection_2d import draw_bounding_boxes
-
 
 class VideoReader(object):
     def __init__(self, file_name):
         self.file_name = file_name
+        self.cap = cv2.VideoCapture(self.file_name)
+        if not self.cap.isOpened():
+            raise IOError('Video {} cannot be opened'.format(self.file_name))
         try:  # OpenCV needs int to read from webcam
             self.file_name = int(file_name)
         except ValueError:
             pass
 
     def __iter__(self):
-        self.cap = cv2.VideoCapture(self.file_name)
         if not self.cap.isOpened():
             raise IOError('Video {} cannot be opened'.format(self.file_name))
         return self
@@ -45,17 +46,26 @@ class VideoReader(object):
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
-    parser.add_argument("--device", help="Device to use (cpu, cuda)", type=str, default="cuda", choices=["cpu", "cuda"])
-    parser.add_argument("--model", help="Model to use", type=str, default="yolov5s6",
-                        choices=['yolov5s', 'yolov5n', 'yolov5m', 'yolov5l', 'yolov5x',
-                                 'yolov5n6', 'yolov5s6', 'yolov5m6', 'yolov5l6', 'custom'])
+    parser.add_argument("--device", help="Device to use (cpu, cuda)", type=str, default="cuda", choices=["cuda", "cpu"])
+    parser.add_argument("--model", help="Model for which a config file will be used", type=str, default="m",
+                        choices=["EfficientNet_Lite0_320", "EfficientNet_Lite1_416", "EfficientNet_Lite2_512",
+                                 "RepVGG_A0_416", "t", "g", "m", "m_416", "m_0.5x", "m_1.5x", "m_1.5x_416",
+                                 "plus_m_320", "plus_m_1.5x_320", "plus_m_416", "plus_m_1.5x_416", "custom"])
+    parser.add_argument("--optimize", help="", type=str, default="", choices=["", "onnx", "jit"])
     args = parser.parse_args()
 
-    yolo_detector = YOLOv5DetectorLearner(model_name=args.model, device=args.device)
+    optimize, device, model = args.optimize, args.device, args.model
+
+    nanodet = NanodetLearner(model_to_use=model, device=device)
+    nanodet.download("./predefined_examples", mode="pretrained")
+    nanodet.load("./predefined_examples/nanodet_{}".format(args.model), verbose=True)
 
     # Use the first camera available on the system
     image_provider = VideoReader(0)
-    fps = -1.0
+
+    if args.optimize != "":
+        nanodet.optimize("./{}/nanodet_{}".format(args.optimize, args.model), optimization=args.optimize, nms_max_num=20)
+
     try:
         counter, avg_fps = 0, 0
         for img in image_provider:
@@ -65,7 +75,7 @@ if __name__ == '__main__':
             start_time = time.perf_counter()
 
             # Perform inference
-            detections = yolo_detector.infer(img)
+            boxes = nanodet.infer(img, conf_threshold=0.35, iou_threshold=0.6, nms_max_num=20)
             end_time = time.perf_counter()
             fps = 1.0 / (end_time - start_time)
 
@@ -74,16 +84,16 @@ if __name__ == '__main__':
 
             img = img.opencv()
 
-            if detections:
-                draw_bounding_boxes(img, detections, yolo_detector.classes, line_thickness=3)
+            if boxes:
+                draw_bounding_boxes(img, boxes, class_names=nanodet.classes)
 
             # Wait a few frames for FPS to stabilize
             if counter > 5:
-                img = cv2.putText(img, "FPS: %.2f" % (avg_fps,), (50, 50), cv2.FONT_HERSHEY_SIMPLEX,
-                                  1, (255, 0, 0), 2, cv2.LINE_AA)
+                image = cv2.putText(img, "FPS: %.2f" % (avg_fps,), (50, 50), cv2.FONT_HERSHEY_SIMPLEX,
+                                    1, (255, 0, 0), 2, cv2.LINE_AA)
 
             cv2.imshow('Result', img)
             cv2.waitKey(1)
             counter += 1
     except:
-        print("Inference fps: ", round(fps))
+        print("Average inference fps: ", avg_fps)